Fix products import: cast unit_price to numeric
COALESCE($5, 0) let Postgres infer the price param as integer (from the 0
literal), so decimal prices failed with 'invalid input syntax for type
integer: 199.99' and every product row was skipped. Cast the param to numeric
so decimals import.

+++ services/api-gateway/src/routes/products-import.ts
@@ -241,7 +241,7 @@ export async function productsImportRoutes(server: FastifyInstance) {
         } else {
           await pool.query(
             `INSERT INTO products (tenant_id, sku, name, description, unit_price, currency, billing_cycle, custom_fields)
-             VALUES ($1, $2, $3, $4, COALESCE($5, 0), COALESCE($6, 'USD'), $7, $8::jsonb)`,
+             VALUES ($1, $2, $3, $4, COALESCE($5::numeric, 0), COALESCE($6, 'USD'), $7, $8::jsonb)`,
             [tenantId, sku, name, description, unitPrice, currency, billing, JSON.stringify(custom)]
           );
           created++;
